requirements.txt
anyio==4.7.0
beautifulsoup4==4.12.3
certifi==2024.8.30
click==8.1.7
dnspython==1.16.0
fastcore==1.7.23
fastlite==0.0.13
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
numpy==2.1.3
oauthlib==3.2.2
packaging==24.2
pandas==2.2.3
pymongo==4.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-fasthtml==0.10.1
python-multipart==0.0.19
pytz==2024.2
PyYAML==6.0.2
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
sqlite-minutils==4.0.3
starlette==0.41.3
typing_extensions==4.12.2
tzdata==2024.2
uvicorn==0.32.1
uvloop==0.21.0
watchfiles==1.0.0
websockets==14.1
